backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
alembic==1.14.1
biopython==1.85
scipy==1.15.1
numpy==2.2.1
psycopg2-binary==2.9.10
python-dotenv==1.0.1
pydantic-settings==2.7.1
httpx==0.28.1
celery[redis]==5.4.0
redis==5.2.1
websockets==14.1
openpyxl==3.1.5
pytest==8.3.4
pytest-asyncio==0.24.0
